Move alarm/operation out of node
